simple rest interface
