+++ api/__init__.py
@@ -7,97 +7,12 @@ from urllib import request
 import redis
 
 from config.bot_config import ICECAST2_SERVERS, SESSION_EXPIRE, ICECAST2_STATS_FILE, SERVERS_LIMIT
+from common import RadioStream, CheckIceCast2Stats
 
-
-class CheckIceCast2Stats:
-
-    """CheckIceCast2Stats"""
-
-    def __init__(self):
-
-        self.servers = []
-
-    def get_stats(self):
-
-        """get_stats"""
-
-        for srv in ICECAST2_SERVERS:
-
-            srv_request = request.urlopen('{}/{}'.format(srv, ICECAST2_STATS_FILE), None, 3)
-
-            if srv_request.getcode() == 200:
-
-                stats_data = json.loads(srv_request.read().decode('utf-8'))
-
-                if 'icestats' in stats_data and 'source' in stats_data['icestats']:
-
-                    # if the server limit reach SERVERS_LIMIT - 5, will not be listen as available
-                    if stats_data['icestats']['source']['listeners'] >= SERVERS_LIMIT - 5:
-                        continue
-
-                    radio = RadioStreams()
-
-                    radio.stream = stats_data['icestats']['source']['listenurl']
-                    radio.server = srv
-                    radio.online = stats_data['icestats']['source']['listeners']
-                    self.servers.append(radio)
-
-        return self.servers
-
-
-class RadioStreams:
-
-    """RadioStreams"""
-
-    def __init__(self):
-
-        """__init__"""
-
-        self.__stream = None
-        self.__server = None
-        self.__online = 0
-
-    @property
-    def stream(self):
-        """stream"""
-        return self.__stream
-
-    @stream.setter
-    def stream(self, stream):
-        """
-        stream
-
-        :param stream
-        """
-        self.__stream = stream
-
-    @property
-    def server(self):
-        """server"""
-        return self.__server
-
-    @server.setter
-    def server(self, host):
-
-        """
-        server
-
-        :param host
-        """
-        self.__server = host
-
-    @property
-    def online(self):
-        """online"""
-        return self.__online
-
-    @online.setter
-    def online(self, online):
-        self.__online = online
 
 class RadioBot:
 
-    """RadioBot"""
+    """ IceCast2 Radio Bot """
 
     def __init__(self):
 
@@ -116,7 +31,12 @@ class RadioBot:
 
         """
 
-        bot.sendMessage(chat_id=update.message.chat_id, text="Hello I'm HMSU Radio Bot.")
+        user_data = bot.get_chat(update.message.chat_id)
+
+        bot.sendMessage(
+            chat_id=update.message.chat_id, text="Hello {}, I'm HMSU Radio Bot.".format(user_data.username)
+        )
+
         bot.sendMessage(chat_id=update.message.chat_id, text="Type /help for full list of commands")
 
     def help_tcb(self, bot, update, args):
@@ -147,7 +67,7 @@ class RadioBot:
 
         user_data = bot.get_chat(update.message.chat_id)
 
-        # if user name is not set, we cannot continue, so we just message back, how to create a username
+        # if username does not exist, we cannot continue, so we just message back, how to create a username
         if not user_data.username:
 
             bot.sendMessage(
@@ -155,7 +75,8 @@ class RadioBot:
             )
 
             bot.sendMessage(
-                chat_id=update.message.chat_id, text="More info @ https://telegram.org/faq#q-what-are-usernames-how-do-i-get-one"
+                chat_id=update.message.chat_id,
+                text="More info @ https://telegram.org/faq#q-what-are-usernames-how-do-i-get-one"
             )
