+++ icecast2/icecast2/urls.py
@@ -13,9 +13,9 @@ Including another URLconf
     1. Import the include() function: from django.conf.urls import url, include
     2. Add a URL to urlpatterns:  url(r'^blog/', include('blog.urls'))
 """
-from django.conf.urls import url
-from django.contrib import admin
+from django.conf.urls import url, include
+from login.views import IceCast2Login
 
 urlpatterns = [
-    url(r'^admin/', admin.site.urls),
+    url(r'^icecast2/', IceCast2Login.as_view()),
 ]
