+++ icecast2/login/views.py
@@ -1,3 +1,9 @@
-from django.shortcuts import render
+from rest_framework import permissions
+from rest_framework.views import APIView
+from rest_framework.response import Response
 
-# Create your views here.
+
+class IceCast2Login(APIView):
+
+    def post(self, request, format=None):
+        return Response({"success": True})
